requirements.txt
streamlit==1.24.0
langchain==0.0.228
openai==0.27.8
python-dotenv==1.0.0
replicate==0.8.0
fpdf==1.7.2
Pillow==9.5.0
